change to long

+++ src/main/java/edu/indiana/soic/spidal/common/sparse/SparseMatrixFile.java
@@ -86,7 +86,7 @@ public class SparseMatrixFile {
             int[] rowPointer = new int[length];
             Arrays.fill(rowPointer, -1);
             int count = 0;
-            int countflips = 0;
+            long countflips = 0;
             //checks if the loop has already completed the row range
             boolean isDone = false;
             Map<Integer, List<double[]>> flipValues = new HashMap<>();
